requirements.txt
streamlit
requests
beautifulsoup4
langchain
python-dotenv
llama-cpp-python
pandas
fastapi
uvicorn[standard]
pydantic
pytest
playwright
tiktoken
tavily-python